requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-stubs==5.0.2
django-stubs-ext==5.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.5.0
tomli==2.0.1
types-PyYAML==6.0.12.20240311
typing_extensions==4.12.2
